+++ src/components/card/TemplateCard.vue
@@ -2,20 +2,23 @@
   <q-card
     v-ripple
     tabindex="0"
-    class="cursor-pointer template-card"
+    class="cursor-pointer template-card transparent"
     :data-cy="`template-card_${template.key}`"
+    flat
   >
-    <q-img
-      :src="`/template-library/templates/${template.key}/icon.svg`"
-      height="100%"
+    <q-card-section class="q-pa-none">
+      <q-img
+        class="rounded-borders template-img bg-white"
+        :src="`/template-library/templates/${template.key}/icon.svg`"
+        fit="fill"
+      />
+    </q-card-section>
+    <q-card-section
+      class="q-px-none q-py-sm text-caption ellipsis-2-lines"
+      data-cy="title-container"
     >
-      <div
-        class="absolute-bottom text-subtitle2 text-center"
-        data-cy="title-container"
-      >
-        {{ template.type }}
-      </div>
-    </q-img>
+      {{ template.type }}
+    </q-card-section>
   </q-card>
 </template>
 
@@ -28,9 +31,15 @@ defineProps({
 });
 </script>
 
-<style scoped>
+<style lang="scss" scoped>
 .template-card {
-  width: 150px;
+  width: 100px;
   height: 150px;
+  overflow: hidden;
+}
+.template-img {
+  width: 100px;
+  height: 100px;
+  border: 1px solid $grey-3;
 }
 </style>